split DmpCore to DmpRunMode DmpDetectorID. Kernel

+++ Event/DmpEvent/include/DmpEvtPMT.h
@@ -28,7 +28,7 @@ private:
 // *  TODO: why can not save all values in this vetor into root file???
 // *
 #pragma message("TODO ----> why can not save all values in this vetor into root file???")
-  std::vector<double>    fADCDynode;        // ADC channels of dynodes
+  std::vector<float>    fADCDynode;        // ADC channels of dynodes
 
   ClassDef(DmpEvtPMT,1)
 };
